Make custom-scripts import in client module compatible and rework case-module content set example of reports to use custom elements
