+++ client/src/index.html
@@ -36,7 +36,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
   <link rel="stylesheet" type="text/css" href="./assets/custom-styles.css">
-  <link rel="stylesheet" type="text/css" href="./assets/custom-scripts.js">
+  <script type="module" src="./assets/custom-scripts.js"></script>
 </head>
 <body>
   <app-root>
