+++ content-sets/case-module/custom-scripts.js
@@ -1,0 +1,2 @@
+import './itemized-report.js'
+import './summarized-report.js'
