+++ content-sets/case-module/reports.html
@@ -1,5 +1,3 @@
-<div style = "padding: 5px;">
-  <h1>Custom Reports</h1>
-  <div id="report1"></div>
-  <div id="report2"></div>
-</div>
+
+<itemized-report></itemized-report>
+<summarized-report></summarized-report>
